pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finance-hub"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "click>=8.1",
    "rich>=13.0",
    "flask>=3.0",
    "python-dateutil>=2.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "black>=24.0",
    "isort>=5.13",
    "mypy>=1.8",
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: fast unit tests with no I/O",
    "integration: tests that hit real DB",
    "slow: tests that take more than a second",
]
